libmemtrack: Use kgsl memory flag to determine usermapped buffers
Instead of using useraddr from kgsl memory allocations to be
matched against proc/<pid>/smaps file to be used to determine
usermapping of a buffer, use the newly added flag which directly
indicates whether a given gpubuffer entry is usermapped or not.
The flag is the last character in the "flags" field.

CRs-fixed: 634962
Change-Id: I1f82f7a2ff207eb780f1938a3b1347451b1e3d77