sdm: handle vbl_pending_ flag in vblank handler.
Since LA uses HandleVSync as vblank_handler & when requesting next vsync event
using RegisterVsync call which sets vbl_pending_ flag as true to indicate that
a vblank event is pending and yet to handle by the vblank_handler.
SetEventState waits for any pending vblank events to be consumed before
disabling the vsync. This change implements the logic in HandleVSync to reset
the vbl_pending_ as there are no pending vblank event.

Change-Id: Icac1c10b60e4b063118c452512b07dd1747fd758
Signed-off-by: Rahul Sharma <sharah@codeaurora.org>