libdrmutils: Fix for copy assignment failure of DRMConnectorInfo
The DRMConnectorInfo object contains non-primitive data type such as
std::string and it was not safe to use the default compiler implemented
copy/move assignment operator for the clearing purpose.
Implemented a clear() member function for DRMConnectorInfo and deleted
the copy assignment operator= in order to block future usage by mistake.

Change-Id: I492f3501b4479c523c957db85650725460b2e189