sdm: drm: Updates related to HDR_RESET state at the end of HDR playback
A new HDR state transition was introduced in the kernel driver to
cover the HDR teardown sequence case and avoid the HDR infoframe
to be programmed repeatedly if there is no change in its contents.

Updating the state machine as:
HDR_ENABLE is used only when starting a HDR playback.
HDR_RESET is used to reset HDR metadata to zero at the end of HDR playback.
This is needed as per the HDMI specification to allow the sink to transition
to traditional gamma from HDR mode.

Change-Id: I0dded291ea222beccc49ed0e885c9fc74b81493d