display-hal: fix no video issue when starting TV app
TV video is not displaying unless pressing other buttons.
This is because if video module did not pass correct color
data to display module, display will use one incorrect CSC
data and program it to hardware. So this patch is to use
one default valid CSC data when such case occurs.

CRs-Fixed: 2331527
Change-Id: I2a4d96264c8ea415168bfd03b9ec4e8bc5a8e3fe
Signed-off-by: Guchun Chen <guchunc@codeaurora.org>