hwc2: fix sideband stream state when cleared by SF unexpectedly
When new sideband stream buffer is received, it allows either SF or
sideband stream thread to commit the buffer and then update the
state. This change fixes the corner case when SF updates before new
buffer flag is set, and clears the state after new buffer flag is
set, which causes a mismatch between new buffer flag and pending
mask flag.

This change also adds timeout when waiting for SF update, in case
if SF only calls validate but not present.

Change-Id: I3ec4bd7b36d381d2b6de1df1e82379d2ec639223