sdm: add NotifierInterface to handle pipe state update event
When SDM is not the only compositor in the system, we need to
synchronize the pipe state between SDM and non-SDM compositor
and transfer pipe resources between SDM compositor and non-SDM
compsitor.

NotifierInterface is added to core interface to send the pipe
state event to SDM core.

HWInfoInterface is moved from libs folder to include folder as
it is required by notifier module to read and update DRM state.

Change-Id: I85f795bdb8b31cbc0380aa4beae67d999c34dd43