sdm: hwc2: Revert "[SDM/HWC2] reduce output fence latency"
This reverts commit bb9f250cf9901653419f544bbf4a0ab75f44eceb.

The original change removes the additional release fence offset
and the latency in hwc. However other parts in sdm doesn't follow
the same design and in the case when GPU composition after HW
composition, invalid fence will be created.

Revert the original change to output the valid fence in all cases.

Change-Id: I536751bab3f7f465b63c6c78aa4bfa10a307c128