sdm: hwc2: clear presentation flag when present is successful
When app layer is updated between validate and present, the present
will fail due to lake of validation, but sideband presentation flag
is still cleared, and as a result the internal dirty flag will be
cleared and cause new layer update not being picked up.

This fix will make sure presentation flag is only cleared when
present is successful.

Change-Id: Id1ce3a96cf655cb24da31cb83c216b3869254959