composer: reset validate if UI layer has scaling
In dual display cases, when the secondary built-in
display has a different resolution than primary
display, it can always have UI scaling. If the panel
is command mode, there can be race condition between
idle power collapse from driver which already resets
the scaler LUT registers before notifying HAL and that
the same time HAL performs commit to driver without
passing the scaler LUTs since it is not needed in
every commit cycle. This causes flicker on the screen
due to the scaler LUTs not configured for this one
commit cycle. Add support to reset validate and pass
scaler LUTs to driver every commit cycle to avoid such
race conditions.

Change-Id: I4bcc7abeb65e8718ca421f8359deacbebb5a47dd